move image functions to shared utilities
