+++ imagedephi/utils/constants.py
